More bits to support >1 minute MQTT intervals
Oops, overlooked that 16 when copy/pasting the macro.

+++ operame.ino
@@ -394,7 +394,7 @@ void setup() {
     if (ota_enabled) setup_ota();
 }
 
-#define every(t) for (static uint16_t _lasttime; (uint16_t)((uint16_t)millis() - _lasttime) >= (t); _lasttime = millis())
+#define every(t) for (static unsigned long _lasttime; (unsigned long)((unsigned long)millis() - _lasttime) >= (t); _lasttime = millis())
 
 void loop() {
     static int co2;
